[521886] Avoid validation in EMFCompareEditingDomain
This change overwrites the creation of the editing domain so that we can
prevent it from doing validation and potential roll-backs.

This change also avoids the problem that the ChangeDescription acts as a
container for removed objects, and this results in UML's CacheAdapter
attaching to the change description, which results in bad performance.
Therefore, we specialize the change description to not attach adapters
at all.

Bug: 521886
Change-Id: If8e518376241225a134705a3a5497422f75d89a2
Signed-off-by: Philip Langer <planger@eclipsesource.com>