Decrease progress monitoring overhead
With the recent changes to add more fine-grained reporting of progress,
it became obvious that the overhead in the UI thread seems high.  A
large comparison with ~330,000 diffs takes almost a minute longer to
open the editor with updating enabled than without due to many events
that are dispatched to the UI thread, beyond what human perception would
appreciate.

Instead we can dispatch a runnable and update that runnable with the
latest change up to the point in time when the UI thread actually
updates the controls. The updates still happen faster than human
perception, but without the overhead of so many actual UI updates. The
implementation involves no locking so has no more overhead on the
calling thread than before.

Change-Id: I7ec3f411f5e520b67178cfba0148a4974694c866
Signed-off-by: Philip Langer <planger@eclipsesource.com>