[525286] [Papyrus] Diffs in invisible stereotype property visualizations not filtered out by default
Pre-requisite for fixing the presentation of diffs for invisible stereotype property visualizations.

The diagram change factory erroneously thinks that a notation view that has its semantic element explicitly set to null inherits the semantic, which it does not.  The applied-stereotype comment views in Papyrus diagrams have explicit null elements, so without this fix, diffs for them are assigned instead to the diagram itself, which (a) doesn't make sense and (b) breaks the filtering to be implemented subsequently.

Bug: 525286
Change-Id: I20493be6a44595e4ca332aeaa786cc012ad3dbd8
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>