Allow loading referenced resources after the comparison is finished
The caching of the NotLoadingResourceSet prevents referenced resources
from being loaded also if the switch allowResourceLoad is set to true.
This switch is used by the EMFCompareStructureMergeViewer to allow
resource loading after the comparison is finished. This would, for
instance, allow to render diagrams, which reference their semantic
models, correctly in the content merge viewer. However, the current
implementation doesn't take this switch into account, if there is a
cached null value already.

With this change, we now take this switch into account when testing
whether we should load resources or use the cache to prevent loading
resources.

Change-Id: I9a30697ee2795dc6bda616fd48cd33944955b841
Signed-off-by: Philip Langer <planger@eclipsesource.com>