EMFCompareStructureMergeViewer does much redundant work changing input
Multiple event bus event are processed, each refreshing the tree, each
selecting the first diff, refreshing the navigatable, refreshing the
title, and so on. We add a guard "inChange" to indicate that we are
processing the input change to avoid all this redundant processing.

Also, the toolbar is enabled by the FetchListener.doneFetching even if
it was previously disabled.

The status control continues to show the last subtask while the view is
updating, which is confusing. Better would be to update to a subtask
indicating that the view is being updated.

In the preference mode where the first diff isn't selected, nothing is
selected, which is kind of confusing and hard to see where focus is.
Better would be to select the first item in the tree.

There are also general problems with focus not ending up in a useful
place.  Better would be to ensure that the tree has focus or if there is
a problem on the problem tab, that the problem tree has focus.

No internalRefresh is needed for a null element.

Change-Id: I9a934cc7a188d0fb1523b921c5390165f7d04476
Signed-off-by: Philip Langer <planger@eclipsesource.com>