The show merge consequences control is confusing when disabled
With the preference to disable the showing of consequences, the tool bar
button does nothing in any case because consequences are never shown.
It's not clear how to enable it again and one might easily conclude
there are no consequences.

Better would be to allow the preview merge mode to be null to indicate
that no consequences should be shown, and better to make this an
explicit choice in the drop down menu.  A new image is needed for that.
Better would be if the menu item style where radio button.

This approach allows the user to disable it on a per-view basis, with
the preference controlling only the initial default for the preview mode
when opening a new editor.  Of course there is a listener to respect the
mode preferences when it changes.

Also, mirroring does not show the expected (correct) results.  It seems
best to invert the mode and of course the computation needs to take into
account the mirroring state as well to show the right inverse things.
The idea being that switching the mirroring state will not show modified
highlighted difference consequence change because the radio choice
changes as well, so the double inverse leads to the same highlighted
results.

Change-Id: I2a72bc9aa465b3e624ab98539a511db259cd3369
Signed-off-by: Philip Langer <planger@eclipsesource.com>