[508526] Improve text compare
The viewer should show the text for the resources of the input (i.e.,
any resource of the logical model) rather than just the resources
initially compared.  It should also properly manage the dirty state of
the overall editor, but also including showing which side is dirty.
Ideally it should be able to show a preview of the resource as it would
be if the user saved with the current state of the processed diffs. It's
also ideal if the objects of the input are actually selected in the
source viewers.

Bug: 508526
Change-Id: Ib35cb9b23399cda672ec38220095d908eba0ce3e
Signed-off-by: Philip Langer <planger@eclipsesource.com>