[527567] DiagramCompareItemProviderAdapterFactory should be a decorator
It shouldn't just extend NotationItemProviderAdapterFactory but then not
properly implement adapters for the whole model. Instead, it should be a
decorator adapter factory and the specialized item providers should be
decorators.  Without this, we won't be able to show properties for the
notation model model because many adapters will be missing or won't be
using the property subclass that knows about all the properties.

We also remove a couple of warnings in the changed classes.

Bug: 527567
Change-Id: I121b09ebc2136d3b237387326f7cfc5cce342931
Signed-off-by: Philip Langer <planger@eclipsesource.com>