[527858] Selecting a diff with conflicts fails
In the content merge viewer, only one of the two conflicting diffs is
associated with the merge viewer item, so if you try to select the other
one, it's not found. We suggest, that if the two diffs are in the same
conflict, only the conflicts need to be compared that they're the same
conflict.

This change also adapts and adds a test specifying that two diffs with
the same conflict in a merge are considered equal but if they are
different diffs are considered unequal.

Bug: 527858
Change-Id: I9e4465082b9710ce89faa7f58a0f8b607f58c59d
Signed-off-by: Philip Langer <planger@eclipsesource.com>