Improve performance of textual value matching
A large fraction of the overhead of creating the textual mergers is the
pointless compilation of regular expression constants. These can be
made static final constants.

Also ThreeWayTextDiff.createBufferedReader converts the string to bytes
and then must convert the bytes back to a string.  This is pointless
overhead, and even error prone, because the system encoding is used and
that might not be able to represent all the characters.  Better to use a
reader that can directly read the existing characters of the string.

Change-Id: I37c47bbddcdaccc9584bd56b8a9df74b6a06f60e
Signed-off-by: Philip Langer <planger@eclipsesource.com>