Improve performance of AbstractCopyCommand.getChangedDiffs
The current algorithm iterates over all object changes, checking each
one if it's one of the relevant diffs; better would be to iterate over
just the relevant diffs, and look each up in the map. This is especially
noticeable if you accept/reject all diffs and then try to change the
state of just one of those diffs (i.e., if there is only one relevant
diff).

Change-Id: I6f87344c7d1730304dbf8485540d138638878148
Signed-off-by: Philip Langer <planger@eclipsesource.com>