[529413] Extension point for defining files included in EMFModelProvider
This change replaces the enablement expression listing all file
extensions explicitly, as used before this change for enabling the
EMFModelProvider for certain files, with a property tester that reads
the files to be included from an extension point instead.

This allows other plug-ins to contribute inclusions of files. These
third-party plug-ins can specify a tester in the form of a content type
or a file extension that shall be handled by the EMF Model Provider, or
in the form of a class that determines whether to include a file or not
at runtime.

The dynamic testers enables third-party plug-ins to perform a more
fine-grained analysis to whether or not include a file, for instance,
Papyrus may not only specify that ALL properties files shall be handled
by the EMFModelProvider but only those where an equally named di file
exists.

Bug: 529413
Change-Id: Id572261fdc4e3b8057b5293338b160230a59815b
Signed-off-by: Philip Langer <planger@eclipsesource.com>