Improve performance of IDiffRelationshipComputer implementations
The current CachingDiffRelationshipComputer implementation directly
implements IDiffRelationshipComputer and wraps an
IDiffRelationshipComputer but that means that calls in that delegate
computer don't end up using the cache. Better would be for
CachingDiffRelationshipComputer to extend DiffRelationshipComputer so
that the caching is used in recursive computations.

Caching the "all" results is not a good idea. Not only are these very
expensive to compute, but caching them all uses O(n^2) memory.
Furthermore, even just computing these values is horribly expensive,
i.e., it can take as much of 3 seconds per diff for
a large model, so imagine processing 100,000 diffs taking 7 days to
complete!

Fortunately, the primary user of these methods is
AbstractMergeRunnable.markAsMerged where it will transition diffs to a
terminal state, so it has no need to process diffs that are already in
the terminal state. As such, two new overrides are introduced to allow
the caller to pass a filter, effectively eliminating this entire cost,
i.e., 100,000 diffs can be processed in 10 seconds instead of 7 days.
AbstractMergeRunnable.HAS_UNRESOLVED_STATE used as the filter.

The method asEmptySet is used to convert what's often an empty set into
a more efficient empty set implementation class; it's significant that
an emptySet()'s iterator is a constant so iteration is cheaper than on a
HashSet that's empty, and of course it's more memory efficient as well.

A computeCache method is added so that a single call to getMerger can be
reused to compute all the cache results for a diff.

The getAllResultingMerges computes many intermediate results, even just
the difference.isEmpty test is O(n) because the filter might
filter out all results. The new implementation computes iterators
instead of so many intermediate sets.

In internalGetAllResultingMerges it's often the case that one of the two
sets is empty, so we can just return one of the two instead of computing
a union.

Finally, in EMFCompareStructureMergeViewer we can compute the cached
information in a background job which effectively improves accept/reject
all performance by 50%.

Change-Id: Id75d693b9b090eed4e50c5d51ff36a253be80a48