Improve IMerger registry implementation
The RegistryImpl maintains a map of registered mergers that must be
repeatedly sorted on each lookup. Better would be to maintain the
registered mergers in a list that is already sorted so that the sorting
cost is paid once per add rather than once per lookup. This way
getMergersByRankDescending can call basicGetMergers which returns an
already-sorted result and even getMergers will return a sorted result,
enough though the API doesn't require that and the old implementation
returned a random/arbitrarily ordered result.

Change-Id: I1ac87627f0f27b5dccef735ac4be40b32148fe53
Signed-off-by: Philip Langer <planger@eclipsesource.com>