Memory leak from the EMFModelProvider.contextToResourceMappingCache
The model provider's cache uses the ResourceMappingContexts as keys,
and these keys are never cleared.

The ISynchronizationContext we're creating here will thus be used as key
and kept in memory. The ISynchronizationScopeManager we're using to
initialize it was an anonymous class and was thus keeping a reference
to its parent class, the editor input, which in turns keeps a reference
to the whole comparison scope as its own comparison input.

Change-Id: Id2dac2d0898cd4003f43e7dec24fe255be42c367