[533467] Fixes "require" linking process in SiriusDiffPostProcessor
The link between the DNode and its 'actualMapping' differences should
not be a "requires" link. Indeed the ReferenceChange for the
'actualMapping' feature "requires" the ReferenceChange of the DNode
addition (default construction). If we add a "requires" link here, we
create a "requires" links cycle. Instead the "implies" links should be
used to force merging the 'actualMapping' ReferenceChange when the
parent difference is merged.

Bug: 533467 
Change-Id: Ifa3c73d50938c4cc83e8e6ff00962c9fefa0e74c