Performance - Computing requirements times
Computing requirements scales quadratically with the number of
differences in the comparison, further amplified when there are a lot
of cluster additions/deletions (adding a container and a cluster of
cascading additions below it).

This aims at filtering as deep down as possible the list of candidate
differences for requirements, and limiting the number of intermediate
collections.

Change-Id: I4d4f7a7cb251e680606edbced946a8baaf7b2325