Prevent CMV disposal when user selects multiple elements in the SMW
This is most likely a Platform/compare bug. The structure merge
viewers don't support multiple selections and if the user selects two
differences, the ContentMergeViewer is disposed and cleared out.

If the user then merges the two differences at once, since no CMV
is currently shown, the editor will not become dirty: the
CompareEditorInput is the one in charge of managing the comparison's
dirty state, and that state is _only_ updated through the listener
(CompareEditorInput#fDirtyStateListener) that is installed on the
opened CMV.

Change-Id: I7c754ca4ce1eb2f6dcf80dd5472c8f787572790d