Ignore "originalTargetObjects" from the EMF Compare change recorder
This is the same performance workaround that has been used by EMF
Transaction in the TransactionChangeRecorder. Each individual "accept"
or "reject" of a difference will trigger the "consolidateChanges" call
on the ChangeRecorder. As the size of the "originalTargetObjects" list
depends on the size of the compared models, creating a copy of this
list every time we merge is a huge performance sink.

Change-Id: Ie597fdbd0d2734963a152c0fdf3099457b1668ef