Add StructureMergeViewer customization options
Improves support for Subclasses of EMFCompareStructureMergeViewer by
changing access modifiers and adding Getters for some private variables.

Although the code is not intended to be subclassed there exist
subclasses with small modifications in the wild. This commit allows
those subclasses to access a little more of the internal mechanism of
the EMFCompareStructureMergeViewer.

Signed-off-by: Stefan Dirix <sdirix@eclipsesource.com>
Change-Id: I7d96b189dd6624a90925d674606a56481681f528