Performance - use DiffCrossReferencer when possible
References with a large amount of elements (5k or more) with a lot of
differences (1k or more) grind the comparison performances to a halt.
When possible, make use of the DiffCrossReferencer to only iterate
over the set of relevant differences instead of the whole set of
differences on the current match.

Change-Id: If80ef7cac5e0e573df78ed902a6b49f2fbcc7211