Cannot merge added/deleted resources
If one of the comparison sides is empty (we're comparing only addition
or deletion of resources with no other resource existing on the
opposite side), EMF Compare is unable to merge the change because
there are no MatchResource allowing us to "find" the opposite
ResourceSet. This is a legacy of the early implementation in which
the comparison scope was no longer available once the comparison was
computed.

Change-Id: Ic1404c984dc651a2d70f53c1f628f673efe629d5