Remove logging from the EMF Compare core
Most of these logs were only start/end information for comparison
phases. This depends on an external logger as the core is standalone.
Log4J 2 does not seem viable for eclipse as it requires an external
configuration file instead of being configurable through API, and we
can't have a global eclipse configuration just for compare. We wish to
remove the dependency towards Log4j 1.

Bug: 578953
Change-Id: Ib421d2f9487f84d8867f936faa0fc068e72e75d3